refactor(slots): extract config_write.py (P3-slots §1f)
Moves the module-level slot-config write/guard pipeline out of
slots/manager.py into slots/config_write.py verbatim: check_npu_exclusivity,
check_default_uniqueness, reconcile_slot_updates,
reconcile_and_guard_slot_config, _reconcile_device_profile,
_base_profile_for_backend, _cfg_effective_backend, plus the private
_read_slot_toml_dict/_iter_peer_configs peer-walk helpers.

Also splits the four leaf cfg accessors (_cfg_to_dict/_cfg_port/
_cfg_provider/_model_default) into a new slots/_cfg_helpers.py so every
upcoming collaborator module (routing/npu/drift/profile_adopt/reaper/
watchdog) can import them without a manager.py import cycle; manager.py
re-exports all of them (and the config_write pipeline) unchanged so
stacks/apply.py's `from hal0.slots.manager import
reconcile_and_guard_slot_config` and slot_view's `_cfg_port` import keep
working verbatim.

No behavior change. tests/slots (379) + tests/stacks (84) green.
